+++ Mnema.API/Content/Scanner.cs
@@ -2,13 +2,8 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Mnema.Models.Entities.Content;
-using Mnema.Models.Publication;
 
 namespace Mnema.API.Content;
-
-public sealed record ParsedTorrentInfo(string Size, List<TorrentFile> Files);
-
-public sealed record TorrentFile(string FileName, string FilePath, long FileSize);
 
 public interface IScannerService
 {
@@ -17,3 +12,7 @@ public interface IScannerService
 
     Task<ParsedTorrentInfo> ParseTorrentFile(string remoteUrl, CancellationToken cancellationToken);
 }
+
+public sealed record ParsedTorrentInfo(string Size, List<TorrentFile> Files);
+
+public sealed record TorrentFile(string FileName, string FilePath, long FileSize);
