+++ Mnema.Frontend/Web/src/app/settings/settings/_components/pages-settings/pages-settings.component.ts
@@ -102,7 +102,6 @@ export class PagesSettingsComponent {
     this.pageService.orderPages(pages.map(p => p.id)).pipe(
       switchMap(() => this.pageService.refreshPages()),
       catchError(err => {
-        this.toastService.genericError(err.error.message);
         this.pages.set(copy);
         return of(pages);
       }),
