+++ Mnema.Providers/Repositories/Nyaa/NyaaRepository.cs
@@ -13,6 +13,7 @@ using Mnema.Common.Helpers;
 using Mnema.Models.DTOs.Content;
 using Mnema.Models.DTOs.UI;
 using Mnema.Models.Entities.Content;
+using Mnema.Providers.Extensions;
 
 namespace Mnema.Providers.Repositories.Nyaa;
